+++ lib/openim_sdk.dart
@@ -51,6 +51,8 @@ export 'src/models/favorite_list_response.dart';
 export 'src/models/favorite_type.dart';
 export 'src/models/link_info.dart';
 export 'src/models/note_info.dart';
+export 'src/models/full_user_info.dart';
+export 'src/models/auth_cache_data.dart';
 
 /// Enums
 export 'src/enums/allow_type.dart';